Renamed EIP712 context files to less generic ones
