+++ src_features/signMessageEIP712/context_712.c
@@ -2,7 +2,7 @@
 
 #include <string.h>
 #include <stdint.h>
-#include "context.h"
+#include "context_712.h"
 #include "mem.h"
 #include "mem_utils.h"
 #include "sol_typenames.h"
