+++ src_features/signMessageEIP712/filtering.c
@@ -4,7 +4,7 @@
 #include "hash_bytes.h"
 #include "ethUstream.h"      // INT256_LENGTH
 #include "apdu_constants.h"  // APDU return codes
-#include "context.h"
+#include "context_712.h"
 #include "commands_712.h"
 #include "typed_data.h"
 #include "path.h"
