+++ src_features/signMessageEIP712/path.c
@@ -4,7 +4,7 @@
 #include <string.h>
 #include "path.h"
 #include "mem.h"
-#include "context.h"
+#include "context_712.h"
 #include "commands_712.h"
 #include "type_hash.h"
 #include "shared_context.h"
